requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pytest>=7.4.0
